+++ features/config-controller.js
@@ -1,4 +1,4 @@
-import { normalizeConfig } from '../runtime/config-service.js?v=20260707-3';
+import { normalizeConfig } from '../runtime/config-service.js?v=20260729-1';
 
 export class ConfigController {
     constructor(dom) {